atualização carrinho

+++ src/app/domain/home/home.component.ts
@@ -33,28 +33,34 @@ export class HomeComponent implements OnInit {
       JSON.parse(localStorage.getItem("produtos")) :
       [];
 
+    // criando um array de itens.
     let item = {
       produto: produto,
       index: produtos.length + 1,
       quantidade: 1
     }
 
-    var teste = true;
+    // Criando uma variável para o item que não existe no carrinho.
+    var itemNaoExistenteNaListaDeProdutos = true;
 
     for (let i = 0; i < produtos.length; i++) {
       if (produtos[i].produto.id == item.produto.id) {
         produtos[i].produto.preco = produtos[i].produto.preco + item.produto.preco;
         produtos[i].quantidade = produtos[i].quantidade + 1;
         localStorage.setItem("produtos", JSON.stringify(produtos));
-        teste = false;
+
+        // Item existente no carrinho, então ele recebe falso.
+        itemNaoExistenteNaListaDeProdutos = false;
       }
     }
 
-    if (teste) {
+    // Caso o item não exista no carrinho, o item será inserido.
+    if (itemNaoExistenteNaListaDeProdutos) {
       produtos.push(item);
+      localStorage.setItem("produtos", JSON.stringify(produtos));
     }
-
-    localStorage.setItem("produtos", JSON.stringify(produtos));
+    
+    // Chamando função que atualiza o número no menu superior
     this.appComponent.atualizaNumero();
   }
